requirements_gcp.txt
google-auth==2.38.0
google-cloud-pubsub==2.28.0
psutil==7.0.0
python-dotenv==1.0.1